refactor: remove separated pagination container
