+++ src/v2/Apps/SavedSearchAlerts/savedSearchAlertsRoutes.tsx
@@ -18,7 +18,7 @@ const OverviewRoute = loadable(
     ),
   {
     resolveComponent: component =>
-      component.SavedSearchAlertsOverviewRouteFragmentContainer,
+      component.SavedSearchAlertsOverviewRoutePaginationContainer,
   }
 )
